Fonts, GreenfootImage
Fully functional font rendering on a GreenfootImage
Shape/color rendering on GreenfootImage

+++ strafer-liberator-master/core/src/com/port/HealthBar.java
@@ -160,8 +160,8 @@ public class HealthBar extends UI {
 
     public void act() {
         if (!moved) {
-            x = getX();
-            y = getY();
+            x = (int) getX();
+            y = (int) getY();
             moved = true;
         }
         if (!WorldData.PAUZA) {
